+++ src/controller/src/controller/account_controller.cpp
@@ -1,22 +1,102 @@
 #include "account_controller.hpp"
 
+#include <qstackedwidget.h>
+
 #include <QAction>
 #include <QMainWindow>
 #include <QObject>
+#include <string>
 
 #include "app/app_context.hpp"
+#include "app/store/account_store.hpp"
 #include "commands/account/create_account_command.hpp"
 #include "commands/undo_stack.hpp"
+#include "config/finance_enums.hpp"
 #include "drafts/account_draft.hpp"
 #include "logging/log_macros.hpp"
+#include "ui/account/account_detail_view.hpp"
 #include "ui/account/create_account_dlg.hpp"
 #include "ui/side_bar/account_category.hpp"
+#include "ui/side_bar/account_item.hpp"
 #include "utils/qt_helpers.hpp"
 
 REGISTER_LOG_CATEGORY("UI.Controller.AccountSideBarController");
 
 namespace controller
 {
+    /**
+     * @brief Controller for managing account-related actions
+     *
+     * @param undoStack A reference to the undo stack, this is used to push
+     * commands that are created as a result of actions in the account category
+     * (e.g. creating a new account), this allows the user to undo and redo
+     * actions related to accounts using the undo stack.
+     * @param appContext A reference to the application context, this is used to
+     * access the stores and services needed to perform operations related to
+     * accounts (e.g. creating a new account), this allows the controller to
+     * interact with the underlying data and business logic for accounts, and
+     * ensures that the controller can perform the necessary operations to
+     * manage accounts effectively.
+     * @param sideBarController A reference to the account side bar controller,
+     * this is used to manage the account selection and display in the side bar,
+     * and allows the controller to respond to account selection events.
+     * @param stackedWidget A pointer to the stacked widget, this is used to
+     * manage the different views for account details and other related
+     * information, allowing the controller to switch between views as needed.
+     */
+    AccountController::AccountController(
+        cmd::UndoStack&           undoStack,
+        app::AppContext&          appContext,
+        AccountSideBarController& sideBarController,
+        QStackedWidget*           stackedWidget
+    )
+        : _undoStack(undoStack),
+          _appContext(appContext),
+          _stackedWidget(stackedWidget),
+          _accountDetailView(new ui::AccountDetailView(_stackedWidget))
+    {
+        connect(
+            &sideBarController,
+            &AccountSideBarController::accountSelected,
+            this,
+            &AccountController::_onAccountSelected
+        );
+
+        _stackedWidget->addWidget(_accountDetailView);
+    }
+
+    /**
+     * @brief Handle the selection of an account in the side bar
+     *
+     * @param id The ID of the selected account
+     */
+    void AccountController::_onAccountSelected(AccountId id)
+    {
+        LOG_ENTRY;
+
+        const auto account =
+            _appContext.getStore().getAccountStore().getAccount(id);
+
+        if (!account.has_value())
+        {
+            LOG_WARNING(
+                std::format("Account with ID {} not found", id.value())
+            );
+            return;
+        }
+
+        const auto accountDraft = drafts::AccountDraft{
+            .id   = std::get<finance::CashAccount>(account.value()).getId(),
+            .name = std::get<finance::CashAccount>(account.value()).getName(),
+            .kind = AccountKind::Cash,
+            .currency =
+                std::get<finance::CashAccount>(account.value()).getCurrency()
+        };
+
+        _accountDetailView->updateAccount(accountDraft);
+        _stackedWidget->setCurrentWidget(_accountDetailView);
+    }
+
     /**
      * @brief Construct a new Account Side Bar Controller:: Account Side Bar
      * Controller object
@@ -171,6 +251,16 @@ namespace controller
         }
 
         refresh();
+    }
+
+    /**
+     * @brief Handle the selection of an account in the side bar
+     *
+     * @param id The ID of the selected account
+     */
+    void AccountSideBarController::onAccountSelected(AccountId id)
+    {
+        emit accountSelected(id);
     }
 
 }   // namespace controller