+++ src/controller/src/controller/account_controller.hpp
@@ -1,12 +1,16 @@
 #ifndef __CONTROLLER__SRC__CONTROLLER__ACCOUNT_CONTROLLER_HPP__
 #define __CONTROLLER__SRC__CONTROLLER__ACCOUNT_CONTROLLER_HPP__
 
-#include "side_bar_category_controller.hpp"
-
-class QAction;       // Forward declaration
-class QMainWindow;   // Forward declaration
+#include <qtmetamacros.h>
 
 #include <QPointer>
+
+#include "config/id_types.hpp"
+#include "side_bar_category_controller.hpp"
+
+class QAction;          // Forward declaration
+class QMainWindow;      // Forward declaration
+class QStackedWidget;   // Forward declaration
 
 namespace drafts
 {
@@ -22,6 +26,7 @@ namespace ui
 {
     class AccountCategory;       // Forward declaration
     class CreateAccountDialog;   // Forward declaration
+    class AccountDetailView;     // Forward declaration
 }   // namespace ui
 
 namespace cmd
@@ -31,12 +36,15 @@ namespace cmd
 
 namespace controller
 {
+
     /**
      * @brief Controller for the account category in the side bar
      *
      */
     class AccountSideBarController : public SideBarCategoryController
     {
+        Q_OBJECT
+
        private:
         /// Reference to the undo stack
         cmd::UndoStack& _undoStack;
@@ -68,10 +76,50 @@ namespace controller
             const QAction*             action
         );
 
+       public slots:
+        void onAccountSelected(AccountId id);
+
+       signals:
+        /**
+         * @brief Signal emitted when an account is selected in the side bar
+         *
+         * @param id The ID of the selected account
+         */
+        void accountSelected(AccountId id);
+
        private slots:
         void _onCreateAccountRequested(const drafts::AccountDraft& account);
     };
 
+    /**
+     * @brief Controller for managing account details
+     *
+     */
+    class AccountController : public QObject
+    {
+        Q_OBJECT
+
+       private:
+        /// Reference to the undo stack
+        cmd::UndoStack& _undoStack;
+        /// Reference to the application context
+        app::AppContext& _appContext;
+        /// Pointer to the stacked widget
+        QStackedWidget* _stackedWidget;
+        /// Pointer to the account detail view
+        QPointer<ui::AccountDetailView> _accountDetailView;
+
+       public:
+        AccountController(
+            cmd::UndoStack&           undoStack,
+            app::AppContext&          appContext,
+            AccountSideBarController& sideBarController,
+            QStackedWidget*           stackedWidget
+        );
+
+       private slots:
+        void _onAccountSelected(AccountId id);
+    };
 }   // namespace controller
 
 #endif   // __CONTROLLER__SRC__CONTROLLER__ACCOUNT_CONTROLLER_HPP__